refactor(l1): move the contact table into the discovery server
The peer table held two unrelated things: the nodes discovery has heard of,
and the peers RLPx is connected to. A consumer that only wants discv5 had to
spin up an actor carrying `PeerConnection`, `Capability` and a `Store`-backed
fork-id filter to get at any of it, and the discv5 tests had to build an
in-memory store to test a nonce counter.

Contacts, k-buckets, the connection pool and discv5 sessions now live in a
`ContactTable` owned outright by `DiscoveryServer` as plain state, so every
discv4/discv5 handler reaches them by `&mut self` instead of paying a message
hop per inbound packet. The peer table keeps only connected peers, their
scores and their request slots.

The two halves meet at five points, all of them now pointing one way, from
RLPx into discovery:

- dial candidates: `next_dial_candidate` returns a `Node`, filtered against a
  connected set discovery maintains itself
- `mark_connected` / `mark_disconnected`, cast alongside the existing
  `new_connected_peer` and `remove_peer` calls
- `set_unwanted` / `set_disposable`, cast from the connection server and the
  peer handler
- lookup pacing, which used to read `target_peers_completion` back out of the
  peer table and is now computed from discovery's own connected count

That last one was the only call pointing from discovery into RLPx; inverting
it is what lets discovery run without a peer table at all. Everything crossing
the boundary is now a cast except the dial-candidate request, so two actors
with sequential mailboxes can never call into each other.

Two fixes fell out of the move:

- `Contact.session` duplicated the standalone session store and was read only
  as a fallback, which outlived the disconnect cleanup: a session was never
  actually dropped for a node that still had a contact. The field is gone and
  the store is the single source of truth.
- `target_reached` was dead, and identical to `target_peers_reached`.
