+++ crates/networking/p2p/discovery/discv5_handlers.rs
@@ -1,4 +1,5 @@
 use crate::{
+    discovery::contact_table::{ContactValidation, DiscoveryProtocol},
     discovery::lookup::{IterativeLookup, LOOKUP_ALPHA, LOOKUP_BUCKET_SIZE},
     discv5::{
         messages::{
@@ -12,7 +13,6 @@ use crate::{
         },
     },
     metrics::METRICS,
-    peer_table::{ContactValidation, DiscoveryProtocol, PeerTableServerProtocol as _},
     rlpx::utils::compress_pubkey,
     types::{Node, NodeRecord},
     utils::{distance, node_id},
@@ -73,11 +73,7 @@ impl DiscoveryServer {
         // (an unauthenticated single-packet DoS of the discv5 actor).
         let src_id = Ordinary::src_id(&packet)?;
 
-        let decrypt_key = self
-            .peer_table
-            .get_session_info(src_id)
-            .await?
-            .map(|s| s.inbound_key);
+        let decrypt_key = self.contacts.session(&src_id).map(|s| s.inbound_key);
 
         let discv5 = self.discv5.as_mut().expect("discv5 state must exist");
 
@@ -169,7 +165,7 @@ impl DiscoveryServer {
             &node.node_id(),
         );
 
-        self.peer_table.set_session_info(node.node_id(), session)?;
+        self.contacts.set_session(node.node_id(), session);
 
         let whoareyou = WhoAreYou::decode(&packet)?;
         let record = (self.local_node_record.seq != whoareyou.enr_seq)
@@ -196,7 +192,7 @@ impl DiscoveryServer {
             DiscoveryServerError::CryptographyError("Invalid ephemeral pubkey".into())
         })?;
 
-        let src_pubkey = if let Some(contact) = self.peer_table.get_contact(src_id).await? {
+        let src_pubkey = if let Some(contact) = self.contacts.get_contact(&src_id) {
             compress_pubkey(contact.node.public_key)
         } else if let Some(record) = &authdata.record {
             if !record.verify_signature() {
@@ -243,7 +239,9 @@ impl DiscoveryServer {
         }
 
         if let Some(record) = &authdata.record {
-            self.peer_table.new_contact_records(vec![record.clone()])?;
+            self.contacts
+                .new_contact_records(vec![record.clone()])
+                .await;
         }
 
         let session = derive_session_keys(
@@ -255,7 +253,7 @@ impl DiscoveryServer {
             false,
         );
 
-        self.peer_table.set_session_info(src_id, session.clone())?;
+        self.contacts.set_session(src_id, session.clone());
         let discv5 = self.discv5.as_mut().expect("discv5 state must exist");
         discv5.session_ips.insert(
             src_id,
@@ -277,12 +275,13 @@ impl DiscoveryServer {
 
     pub(crate) async fn discv5_revalidate(&mut self) -> Result<(), DiscoveryServerError> {
         if let Some(contact) = self
-            .peer_table
-            .get_contact_to_revalidate(REVALIDATION_INTERVAL, DiscoveryProtocol::Discv5)
-            .await?
-            && let Err(e) = self.discv5_send_ping(&contact.node).await
+            .contacts
+            .contact_to_revalidate(REVALIDATION_INTERVAL, DiscoveryProtocol::Discv5)
         {
-            trace!(protocol = "discv5", node = %contact.node.node_id(), err = ?e, "Failed to send revalidation PING");
+            let node = contact.node.clone();
+            if let Err(e) = self.discv5_send_ping(&node).await {
+                trace!(protocol = "discv5", node = %node.node_id(), err = ?e, "Failed to send revalidation PING");
+            }
         }
         Ok(())
     }
@@ -318,9 +317,8 @@ impl DiscoveryServer {
 
         // Seed with closest known nodes from the connection pool
         let seed = self
-            .peer_table
-            .get_closest_from_pool(target_id, LOOKUP_BUCKET_SIZE)
-            .await?;
+            .contacts
+            .closest_from_pool(target_id, LOOKUP_BUCKET_SIZE);
         if seed.is_empty() {
             trace!(
                 protocol = "discv5",
@@ -365,7 +363,7 @@ impl DiscoveryServer {
             let find_node_msg = self.discv5_build_find_node_for_target(target, &node);
             if let Err(e) = self.discv5_send_ordinary(find_node_msg, &node).await {
                 debug!(protocol = "discv5", sending = "FindNode", addr = ?node.udp_addr(), err=?e, "Error sending message");
-                self.peer_table.set_disposable(node_id)?;
+                self.contacts.set_disposable(&node_id);
                 METRICS.record_new_discarded_node();
                 if let Some(discv5) = &mut self.discv5
                     && let Some(lookup) = discv5.active_lookups.get_mut(idx)
@@ -411,9 +409,12 @@ impl DiscoveryServer {
         });
 
         if outbound_key.is_none()
-            && let Some(contact) = self.peer_table.get_contact(sender_id).await?
+            && let Some(node) = self
+                .contacts
+                .get_contact(&sender_id)
+                .map(|c| c.node.clone())
         {
-            return self.discv5_send_ordinary(pong, &contact.node).await;
+            return self.discv5_send_ordinary(pong, &node).await;
         }
         let key = self
             .discv5_resolve_outbound_key(&sender_id, outbound_key)
@@ -429,25 +430,29 @@ impl DiscoveryServer {
         pong_message: PongMessage,
         sender_id: H256,
     ) -> Result<(), DiscoveryServerError> {
-        self.peer_table
-            .record_pong_received(sender_id, pong_message.req_id)?;
+        self.contacts
+            .record_pong_received(&sender_id, &pong_message.req_id);
 
-        if let Some(contact) = self.peer_table.get_contact(sender_id).await? {
-            let cached_seq = contact.record.as_ref().map_or(0, |r| r.seq);
-            if pong_message.enr_seq > cached_seq {
-                trace!(
-                    protocol = "discv5",
-                    from = %sender_id,
-                    cached_seq,
-                    pong_seq = pong_message.enr_seq,
-                    "ENR seq mismatch, requesting updated ENR (FINDNODE distance 0)"
-                );
-                let find_node = Message::FindNode(FindNodeMessage {
-                    req_id: generate_req_id(),
-                    distances: vec![0],
-                });
-                self.discv5_send_ordinary(find_node, &contact.node).await?;
-            }
+        // Copied out rather than held: the table is plain state, so a live
+        // `&Contact` would block the `&mut self` send below.
+        if let Some((node, cached_seq)) = self
+            .contacts
+            .get_contact(&sender_id)
+            .map(|c| (c.node.clone(), c.record.as_ref().map_or(0, |r| r.seq)))
+            && pong_message.enr_seq > cached_seq
+        {
+            trace!(
+                protocol = "discv5",
+                from = %sender_id,
+                cached_seq,
+                pong_seq = pong_message.enr_seq,
+                "ENR seq mismatch, requesting updated ENR (FINDNODE distance 0)"
+            );
+            let find_node = Message::FindNode(FindNodeMessage {
+                req_id: generate_req_id(),
+                distances: vec![0],
+            });
+            self.discv5_send_ordinary(find_node, &node).await?;
         }
 
         let discv5 = self.discv5.as_mut().expect("discv5 state must exist");
@@ -478,11 +483,7 @@ impl DiscoveryServer {
         sender_addr: SocketAddr,
         outbound_key: Option<[u8; 16]>,
     ) -> Result<(), DiscoveryServerError> {
-        let send_to_contact = match self
-            .peer_table
-            .validate_contact(sender_id, sender_addr.ip())
-            .await?
-        {
+        let send_to_contact = match self.contacts.validate_contact(sender_id, sender_addr.ip()) {
             ContactValidation::Valid(contact) => Some(*contact),
             ContactValidation::UnknownContact => None,
             reason => {
@@ -492,9 +493,8 @@ impl DiscoveryServer {
         };
 
         let mut nodes = self
-            .peer_table
-            .get_nodes_at_distances(find_node_message.distances.clone())
-            .await?;
+            .contacts
+            .nodes_at_distances(&find_node_message.distances);
         if find_node_message.distances.contains(&0) {
             nodes.push(self.local_node_record.clone());
         }
@@ -541,8 +541,9 @@ impl DiscoveryServer {
         &mut self,
         nodes_message: NodesMessage,
     ) -> Result<(), DiscoveryServerError> {
-        self.peer_table
-            .new_contact_records(nodes_message.nodes.clone())?;
+        self.contacts
+            .new_contact_records(nodes_message.nodes.clone())
+            .await;
 
         // Feed results into ALL active lookups (but don't advance — the timer
         // drives lookup progress so that traffic stays controlled).
@@ -572,7 +573,7 @@ impl DiscoveryServer {
         });
 
         self.discv5_send_ordinary(ping, node).await?;
-        self.peer_table.record_ping_sent(node.node_id(), req_id)?;
+        self.contacts.record_ping_sent(&node.node_id(), req_id);
 
         Ok(())
     }
@@ -591,7 +592,7 @@ impl DiscoveryServer {
             src_id: self.local_node.node_id(),
             message: message.clone(),
         };
-        let encrypt_key = match self.peer_table.get_session_info(node.node_id()).await? {
+        let encrypt_key = match self.contacts.session(&node.node_id()) {
             Some(s) => s.outbound_key,
             None => {
                 trace!(
@@ -620,14 +621,14 @@ impl DiscoveryServer {
     }
 
     async fn discv5_resolve_outbound_key(
-        &self,
+        &mut self,
         node_id: &H256,
         key: Option<[u8; 16]>,
     ) -> Result<[u8; 16], DiscoveryServerError> {
         if let Some(key) = key {
             return Ok(key);
         }
-        match self.peer_table.get_session_info(*node_id).await? {
+        match self.contacts.session(node_id) {
             Some(s) => Ok(s.outbound_key),
             None => {
                 trace!(
@@ -688,7 +689,7 @@ impl DiscoveryServer {
             record,
             message: message.clone(),
         };
-        let encrypt_key = match self.peer_table.get_session_info(node.node_id()).await? {
+        let encrypt_key = match self.contacts.session(&node.node_id()) {
             Some(s) => s.outbound_key,
             None => {
                 trace!(
@@ -781,9 +782,8 @@ impl DiscoveryServer {
         let mut rng = OsRng;
 
         let enr_seq = self
-            .peer_table
-            .get_contact(src_id)
-            .await?
+            .contacts
+            .get_contact(&src_id)
             .map_or(0, |c| c.record.as_ref().map_or(0, |r| r.seq));
 
         let who_are_you = WhoAreYou {
@@ -815,7 +815,7 @@ impl DiscoveryServer {
     }
 
     async fn discv5_send_packet(
-        &self,
+        &mut self,
         packet: &Packet,
         dest_id: &H256,
         addr: SocketAddr,
