+++ crates/networking/p2p/network.rs
@@ -4,8 +4,9 @@ use crate::rlpx::l2::l2_connection::P2PBasedContext;
 #[derive(Clone, Debug)]
 pub struct P2PBasedContext;
 use crate::{
-    discovery::{DiscoveryConfig, DiscoveryServer, DiscoveryServerError},
+    discovery::{DiscoveryConfig, DiscoveryHandle, DiscoveryServer, DiscoveryServerError},
     metrics::{CurrentStepValue, METRICS},
+    peer_filter::EthForkIdFilter,
     peer_table::{PeerData, PeerTable, PeerTableServerProtocol as _},
     rlpx::{
         connection::server::{PeerConnBroadcastSender, PeerConnection},
@@ -37,6 +38,9 @@ pub struct P2PContext {
     pub tracker: TaskTracker,
     pub signer: SecretKey,
     pub table: PeerTable,
+    /// Reaches the discovery server, once one is running. Inert when p2p runs
+    /// without discovery.
+    pub discovery: DiscoveryHandle,
     pub storage: Store,
     pub blockchain: Arc<Blockchain>,
     pub(crate) broadcast: PeerConnBroadcastSender,
@@ -96,6 +100,7 @@ impl P2PContext {
             tracker,
             signer,
             table: peer_table,
+            discovery: DiscoveryHandle::new(),
             storage,
             blockchain,
             broadcast: channel_broadcast_send_end,
@@ -152,12 +157,12 @@ pub async fn start_network(
 
     let local_node_record = build_local_node_record(&context).await?;
 
-    DiscoveryServer::spawn(
+    let discovery = DiscoveryServer::spawn(
         context.local_node.clone(),
         local_node_record,
         context.signer,
         udp_socket,
-        context.table.clone(),
+        Box::new(EthForkIdFilter::new(context.storage.clone())),
         bootnodes,
         config,
     )
@@ -165,6 +170,12 @@ pub async fn start_network(
     .inspect_err(|e| {
         error!("Failed to start discovery server: {e}");
     })?;
+
+    // Publishes the server to every clone of the context, which is what lets the
+    // initiator ask for dial candidates and connections report their lifecycle.
+    if !context.discovery.set(discovery) {
+        error!("Discovery server was already set on this context");
+    }
 
     context.tracker.spawn(serve_p2p_requests(context.clone()));
