+++ crates/networking/p2p/rlpx/connection/handshake.rs
@@ -145,6 +145,7 @@ pub(crate) async fn perform(
             client_version: context.client_version.clone(),
             connection_broadcast_send: context.broadcast.clone(),
             peer_table: context.table.clone(),
+            discovery: context.discovery.clone(),
             #[cfg(feature = "l2")]
             l2_state: context
                 .based_context
